requirement.txt
alembic==1.7.7
click==8.1.0
colorama==0.4.4
Flask==2.1.0
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.3
itsdangerous==2.1.2
Jinja2==3.1.1
Mako==1.2.0
MarkupSafe==2.1.1
numpy==1.22.3
oracledb==1.4.1
SQLAlchemy==1.4.32
typing-extensions==4.1.1
Werkzeug==2.0.3
zipp==3.7.0
psycopg2-binary==2.9.9
Flask_Cors==5.0.0
